Refactor: Organize project into packages
This refactoring organizes the project's source code into a more structured package hierarchy. All Java classes were moved from the default `com.tuempresa.proyecto.demo1` package into sub-packages based on their functionality.

The new package structure is as follows:
- `game`: Core game logic and configuration.
- `model`: Data model classes (e.g., `Snake`, `GameState`).
- `net`: Networking classes for client-server communication.
- `net.dto`: Data Transfer Objects for network communication.
- `ui`: Swing-based user interface components.
- `util`: Utility classes like the logger.
- `performance`: Tests related to performance metrics.

All `package` and `import` statements have been updated accordingly to ensure the project compiles and all tests pass. An unused FXML file was also removed.
