--- src/main/java/com/tuempresa/proyecto/demo1/BotDifficulty.java
@@ -1,7 +1,0 @@
-package com.tuempresa.proyecto.demo1;
-
-public enum BotDifficulty {
-    FACIL,
-    INTERMEDIO,
-    MAESTRO
-}
