--- src/main/java/com/tuempresa/proyecto/demo1/Direccion.java
@@ -1,9 +1,0 @@
-package com.tuempresa.proyecto.demo1;
-
-public enum Direccion {
-    ARRIBA,
-    ABAJO,
-    IZQUIERDA,
-    DERECHA,
-    NINGUNA // Útil para cuando no hay input
-}
