--- src/main/java/com/tuempresa/proyecto/demo1/GamePhase.java
@@ -1,7 +1,0 @@
-package com.tuempresa.proyecto.demo1;
-
-public enum GamePhase {
-    WAITING_FOR_PLAYERS,
-    IN_PROGRESS,
-    GAME_ENDED
-}
