--- src/main/java/com/tuempresa/proyecto/demo1/GameStateSnapshot.java
@@ -1,18 +1,0 @@
-package com.tuempresa.proyecto.demo1;
-
-public class GameStateSnapshot implements java.io.Serializable {
-    private static final long serialVersionUID = 1L;
-    public final int width, height;
-    public final java.util.List<SnakeSnapshot> snakes;
-    public final java.util.List<FrutaSnapshot> frutas;
-    public final boolean juegoActivo;
-    public final GamePhase gamePhase; // <-- NUEVO
-    public GameStateSnapshot(int width, int height, java.util.List<SnakeSnapshot> snakes, java.util.List<FrutaSnapshot> frutas, boolean juegoActivo, GamePhase gamePhase) {
-        this.width = width;
-        this.height = height;
-        this.snakes = java.util.Collections.unmodifiableList(new java.util.ArrayList<>(snakes));
-        this.frutas = java.util.Collections.unmodifiableList(new java.util.ArrayList<>(frutas));
-        this.juegoActivo = juegoActivo;
-        this.gamePhase = gamePhase; // <-- NUEVO
-    }
-}
