+++ src/main/java/com/tuempresa/proyecto/demo1/model/BotDifficulty.java
@@ -1,0 +1,7 @@
+package com.tuempresa.proyecto.demo1.model;
+
+public enum BotDifficulty {
+    FACIL,
+    INTERMEDIO,
+    MAESTRO
+}
