+++ src/main/java/com/tuempresa/proyecto/demo1/model/Direccion.java
@@ -1,0 +1,9 @@
+package com.tuempresa.proyecto.demo1.model;
+
+public enum Direccion {
+    ARRIBA,
+    ABAJO,
+    IZQUIERDA,
+    DERECHA,
+    NINGUNA // Útil para cuando no hay input
+}
