+++ src/main/java/com/tuempresa/proyecto/demo1/model/GamePhase.java
@@ -1,0 +1,7 @@
+package com.tuempresa.proyecto.demo1.model;
+
+public enum GamePhase {
+    WAITING_FOR_PLAYERS,
+    IN_PROGRESS,
+    GAME_ENDED
+}
